+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 8.3
 -----
+- Add Deselect action from long press [#3786](https://github.com/Automattic/pocket-casts-ios/pull/3786)
 - Fix WatchManager to only report unknown message if it was truly unknown / unprocessed [#3748](https://github.com/Automattic/pocket-casts-ios/issues/3748)
 
 8.2
 -----
+- Playlists: introduce DiffableKit to handle cell reloading [#3784](https://github.com/Automattic/pocket-casts-ios/pull/3784)
+- Playlists: improve artwork and counts queries [#3783](https://github.com/Automattic/pocket-casts-ios/pull/3783), [#3807](https://github.com/Automattic/pocket-casts-ios/pull/3807)
+- Improve app performance when trigering multiple downloads [#3809](https://github.com/Automattic/pocket-casts-ios/pull/3809)
 
 8.1.1
 -----